refactor: smart-trigger to typescript
