+++ reports/smart-trigger-report.json
@@ -1,0 +1,26 @@
+{
+  "timestamp": "2025-09-20T18:22:17.328Z",
+  "changes": {
+    "contracts": 0,
+    "functions": [],
+    "security": 0,
+    "dependencies": 1,
+    "config": 1
+  },
+  "triggers": {
+    "fullSecurityScan": true,
+    "deepContractAnalysis": false,
+    "performanceTests": false,
+    "integrationTests": true,
+    "deploymentTests": true,
+    "gasTests": false,
+    "auditTests": false
+  },
+  "riskScore": 20,
+  "riskLevel": "medium",
+  "summary": {
+    "totalFilesChanged": 8,
+    "triggeredScans": 3,
+    "triggerEfficiency": 43
+  }
+}
